Support for maps
This change adds support for maps which includes the sub-types
of certificates and rsa keys. It is not general purpose map
support. Since this change will also add a "guess" of the data
type rather than just stating "value" a note about double-checking
the output is added.

The code has a bit of a hack. Since the format of the input
can vary widely we use a poor-man's union by allowing that an Importable
can contain either a value directly as a string or a map. The YAML
library will now allow us to use the name 'value' for the same field,
even though only one would appear at a time. This means that we have to
set a different element name for maps - and then change the name before
printing the results. This is a bit hacky but is significantly simpler
than hand-writing our own marshal/unmarshal code which would then have
to be maintained and tested.